Run modelMethods on toJSON output not attributes
The best way I've found to remove a password from a user object is via
the `omit` method.  This matches up with the documentation that shows
the following example:

``` javascript
Customer.login(email, password).then(function(customer) {
  res.json(customer.omit('password'));
})
```

This works great until you use `withRelated` to fetch a user profile,
for instance.  `omit` will run on the `attributes` not the merged
output, so using the above example would strip out the fetched
additionals.

This PR is my naive solution at fixing the problem.  I updated the base
model to use `toJSON` as well as the virtuals plugin.

Let me know if anything else needs documenting or if is avoidable
intentional behavior.