Fix two small typos.
I noticed two small grammatical errors near the top of the front page while reading through the Bookshelf docs. Since these are on the main page and close to the top, I thought they might be worth correcting. Both should increase the readability of the paragraph in which they appear.